Setup development
