+++ .gitignore
@@ -62,3 +62,6 @@ typings/
 
 # TypeScript build output
 build/*
+
+# Development environments
+.vscode
