+++ package.json
@@ -2,17 +2,19 @@
   "name": "kubercade",
   "version": "1.0.0",
   "description": "Serverless containerized social arcade built on Knative.",
-  "main": "build/index.js",
+  "main": "build/app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "prod": "tsc && node ./build/index.js",
+    "dev": "ts-node-dev --respawn --transpileOnly ./src/app.ts",
+    "prod": "tsc && node ./build/app.js",
     "check": "gts check",
     "clean": "gts clean",
     "compile": "tsc -p .",
     "fix": "gts fix",
     "prepare": "npm run compile",
     "pretest": "npm run compile",
-    "posttest": "npm run check"
+    "posttest": "npm run check",
+    "test": "mocha --exit -r ts-node/register tests/**/*.test.ts",
+    "testWithCoverage": "nyc -r lcov -e .ts -x \"*.test.ts\" mocha -r ts-node/register tests/**/*.test.ts && nyc report"
   },
   "repository": {
     "type": "git",
@@ -28,18 +30,22 @@
   },
   "homepage": "https://github.com/knative-portability/Kubercade#readme",
   "dependencies": {
-    "chai": "^4.2.0",
     "express": "^4.17.1",
-    "mocha": "^6.2.0",
-    "sinon": "^7.4.1",
-    "typescript": "^3.5.3"
+    "sinon": "^7.4.1"
   },
   "devDependencies": {
+    "@types/chai": "^4.2.0",
     "@types/express": "^4.17.0",
+    "@types/mocha": "^5.2.7",
     "@types/node": "^10.0.3",
+    "chai": "^4.2.0",
+    "chai-http": "^4.3.0",
+    "gts": "^1.1.0",
+    "mocha": "^6.2.0",
+    "nyc": "^14.1.1",
+    "ts-node": "^8.3.0",
     "ts-node-dev": "^1.0.0-pre.40",
     "tslint": "^5.18.0",
-    "gts": "^1.1.0",
-    "typescript": "~3.5.0"
+    "typescript": "^3.5.3"
   }
 }
