+++ tsconfig.json
@@ -3,7 +3,8 @@
   "compilerOptions": {
     "rootDir": ".",
     "outDir": "build",
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "lib": ["dom"]
   },
   "include": [
     "src/**/*.ts",
